# docs/api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Atlas API",
    "version": "v1"
  },
  "paths": {
...
    "/onboarding/start": {
      "post": {
        "tags": [
          "Onboarding"
        ],
        "summary": "Start onboarding: creates Tenant + User (HostAdmin) + draft Property + draft Listing.\r\nReturns JWT + created IDs + next step.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Atlas.Api.DTOs.OnboardingStartRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Atlas.Api.DTOs.OnboardingStartRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Atlas.Api.DTOs.OnboardingStartRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Atlas.Api.DTOs.OnboardingStartResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/onboarding/profile": {
      "put": {
        "tags": [
          "Onboarding"
        ],
        "summary": "Update tenant profile (legal/tax details).",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Atlas.Api.DTOs.OnboardingProfileUpdateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Atlas.Api.DTOs.OnboardingProfileUpdateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Atlas.Api.DTOs.OnboardingProfileUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/onboarding/documents": {
      "post": {
        "tags": [
          "Onboarding"
        ],
        "summary": "Upload KYC document (multipart form).",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "docType": {
                    "type": "string"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "docType": {
                  "style": "form"
                },
                "file": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Atlas.Api.DTOs.HostKycDocumentDto"
                }
              }
            }
          }
        }
      }
    },
    "/onboarding/airbnb/prefill": {
      "post": {
        "tags": [
          "Onboarding"
        ],
        "summary": "Prefill property info from Airbnb URL or pasted text (public metadata only).",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Atlas.Api.DTOs.AirbnbPrefillRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Atlas.Api.DTOs.AirbnbPrefillRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Atlas.Api.DTOs.AirbnbPrefillRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Atlas.Api.DTOs.AirbnbPrefillResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/onboarding/publish": {
      "post": {
        "tags": [
          "Onboarding"
        ],
        "summary": "Publish: validates PublishGate checklist, sets listing + tenant active.\r\nReturns 200 on success, 422 with blockers if not ready.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Atlas.Api.DTOs.AirbnbPrefillRequestDto": {
        "type": "object",
        "properties": {
          "airbnbUrl": {
            "maxLength": 500,
            "type": "string",
            "nullable": true
          },
          "pastedText": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Atlas.Api.DTOs.AirbnbPrefillResponseDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "locationText": {
            "type": "string",
            "nullable": true
          },
          "propertyType": {
            "type": "string",
            "nullable": true
          },
          "amenities": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "houseRules": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "maxGuests": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "source": {
            "type": "string",
            "nullable": true
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Atlas.Api.DTOs.HostKycDocumentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "docType": {
            "type": "string",
            "nullable": true
          },
          "fileUrl": {
            "type": "string",
            "nullable": true
          },
          "originalFileName": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "Atlas.Api.DTOs.OnboardingProfileUpdateDto": {
        "type": "object",
        "properties": {
          "legalName": {
            "maxLength": 200,
            "type": "string",
            "nullable": true
          },
          "displayName": {
            "maxLength": 200,
            "type": "string",
            "nullable": true
          },
          "businessType": {
            "maxLength": 30,
            "type": "string",
            "nullable": true
          },
          "registeredAddressLine": {
            "maxLength": 500,
            "type": "string",
            "nullable": true
          },
          "city": {
            "maxLength": 100,
            "type": "string",
            "nullable": true
          },
          "state": {
            "maxLength": 50,
            "type": "string",
            "nullable": true
          },
          "pincode": {
            "maxLength": 10,
            "type": "string",
            "nullable": true
          },
          "pan": {
            "pattern": "^[A-Z]{5}[0-9]{4}[A-Z]$",
            "type": "string",
            "description": "Full PAN (ABCDE1234F). Stored as hash + last4 only.",
            "nullable": true
          },
          "gstin": {
            "maxLength": 15,
            "type": "string",
            "nullable": true
          },
          "placeOfSupplyState": {
            "maxLength": 50,
            "type": "string",
            "nullable": true
          },
          "primaryEmail": {
            "maxLength": 200,
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "primaryPhone": {
            "maxLength": 20,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Atlas.Api.DTOs.OnboardingStartRequestDto": {
        "required": [
          "displayName",
          "email",
          "password",
          "phone"
        ],
        "type": "object",
        "properties": {
          "email": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string",
            "format": "email"
          },
          "phone": {
            "maxLength": 20,
            "minLength": 1,
            "type": "string"
          },
          "displayName": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "password": {
            "minLength": 6,
            "type": "string"
          },
          "city": {
            "maxLength": 100,
            "type": "string",
            "nullable": true
          },
          "pincode": {
            "maxLength": 10,
            "type": "string",
            "nullable": true
          },
          "propertyType": {
            "maxLength": 30,
            "type": "string",
            "nullable": true
          },
          "roomCount": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "addressLine": {
            "maxLength": 500,
            "type": "string",
            "nullable": true
          },
          "airbnbUrl": {
            "maxLength": 500,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Atlas.Api.DTOs.OnboardingStartResponseDto": {
        "type": "object",
        "properties": {
          "tenantSlug": {
            "type": "string",
            "nullable": true
          },
          "tenantId": {
            "type": "integer",
            "format": "int32"
          },
          "propertyId": {
            "type": "integer",
            "format": "int32"
          },
          "listingId": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "token": {
            "type": "string",
            "nullable": true
          },
          "nextStep": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Microsoft.AspNetCore.Mvc.ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      }
    },
    "securitySchemes": {
      "X-Tenant-Slug": {
        "type": "apiKey",
        "description": "Tenant slug (e.g. atlas). Required in Production.",
        "name": "X-Tenant-Slug",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "X-Tenant-Slug": [ ]
    }
  ]
}